Создает компонент для пустого списка предложений на главной странице, а основной контент переносит в компонент Places
